feat(trpg): move common commands into extension
