+++ dice/ext_core.go
@@ -61,9 +61,5 @@ func RegisterBuiltinExtCore(dice *Dice) {
 		},
 	}
 
-	theExt.CmdMap = map[string]*CmdItemInfo{
-		"team": cmdTeam,
-	}
-
 	dice.RegisterExtension(theExt)
 }
