+++ dice/ext_trpg_team.go
