+++ dice/ext_trpg_team_test.go
