+++ docs/trpg.md
@@ -1,8 +1,8 @@
-# `trpg.st` 与人物卡 Hook
+# `trpg` 通用跑团扩展
 
 ## 定位与兼容性
 
-`trpg` 是一个默认不激活的内置扩展，提供独立于 `coc7.st` 和 `dnd5e.st` 的通用 `.st` 实现。新规则模板可以这样选择它：
+`trpg` 是默认激活的内置扩展，集中提供通用 `.st`、`.sn` 和 `.team` 指令。其中 `.st` 拥有独立于 `coc7.st` 和 `dnd5e.st` 的事务与 Hook 实现。新规则模板可以这样选择它：
 
 ```yaml
 set:
@@ -11,14 +11,14 @@ set:
     - my-rule
 ```
 
-该设计不会替换已有指令：
+指令提供者继续遵循现有的 `relatedExt` 优先级：
 
 - 仍声明 `relatedExt: [coc7, ...]` 的模板继续使用 `coc7.st`；
 - DND5E 模板继续使用 `dnd5e.st`；
-- 只有显式关联并激活 `trpg` 的模板才使用 `trpg.st`；
+- 声明 `relatedExt: [trpg, ...]` 的模板优先使用 `trpg.st`；
 - `trpg.st` 未注册任何 hook 时，属性解析和数据变更行为与通用 `coc7.st` 保持一致；回复使用独立的 `TRPG:*` 文本模板。
 
-扩展选择和顺序仍遵循现有的 `relatedExt` 指令优先级机制。系统不会额外检测多个 `.st` 提供者之间的冲突。
+`.sn` 与 `.team` 的既有实现和数据结构完整迁移至 `trpg`。由于 `trpg` 默认激活，常规使用方式保持不变。系统不会额外检测多个 `.st` 提供者之间的冲突。
 
 ## 注册
